backend/requirements.txt
beautifulsoup4==4.11.1
Django==3.2.16
djangorestframework==3.14.0
requests==2.28.1
django-allauth==0.52.0
django-cors-headers==3.13.0
Pillow==9.2.0
lxml==4.9.1